Linestyle improvements from 3.0
Reviewers: staniek

Tags: #kreport

Differential Revision: https://phabricator.kde.org/D5404

+++ src/common/KReportItemLine.h
@@ -54,8 +54,8 @@ protected:
     KProperty *m_lineStyle;
 
     KReportLineStyle lineStyle() const;
-    int weight() const;
-    void setWeight(int w);
+    qreal weight() const;
+    void setWeight(qreal w);
 
 private:
     virtual void createProperties();
